requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1
plotly==5.24.1
yfinance==0.2.51
firebase-admin==6.5.0
pyrebase4==4.8.0
pydantic==2.10.4
python-dotenv==1.0.1
